Rename class and apply some first best practices
